+++ nucliadb_utils/nucliadb_utils/settings.py
@@ -22,22 +22,7 @@ from typing import Dict, List, Optional
 from pydantic import BaseSettings
 from pydantic.class_validators import root_validator
 
-
-class RunningSettings(BaseSettings):
-    debug: bool = True
-    sentry_url: Optional[str] = None
-    running_environment: str = "local"
-    logging_integration: bool = False
-    log_level: str = "DEBUG"
-    activity_log_level: str = "INFO"
-    swim_level: str = "INFO"
-    metrics_port: int = 3030
-    metrics_host: str = "0.0.0.0"
-    serving_port: int = 8080
-    serving_host: str = "0.0.0.0"
-
-
-running_settings = RunningSettings()
+from nucliadb_utils.settings_running import running_settings  # noqa
 
 
 class HTTPSettings(BaseSettings):