+++ nucliadb_utils/CHANGELOG.rst
@@ -4,7 +4,7 @@ CHANGELOG
 1.0.6 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Split running settings on a different module
 
 
 1.0.5 (2022-05-31)